pyproject.toml
[project]
name = "timestamper"
version = "0.1.0"
description = "A PySide6 application for adding and editing EXIF data to images."
keywords = ["exif", "pyside6", "image", "metadata", "timestamp"]
authors = [
    { name = "Logan Wu", email = "loganbwu@gmail.com" }
]
dependencies = [
    "pyexiftool==0.5.6",
    "pyside6==6.6.1",
    "pyside6-addons==6.6.1",
    "pyside6-essentials==6.6.1",
    "shiboken6==6.6.1",
    "pytest>=8.4.1",
    "pytest-qt>=4.5.0",
    "numpy==1.26.4",
]
readme = "README.md"
requires-python = ">= 3.12"

[project.scripts]
timestamper = "timestamper.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/timestamper"]

[tool.pytest.ini_options]
pythonpath = ["."]
